refactor: 클럽 검색을 기존 ClubController에서 분리 - ClubController -> ClubProfileController, ClubSearchController - ClubCommandService -> ClubProfileService - remove ClubDetailedPageService
