--- backend/src/main/java/moadong/club/service/ClubDetailedPageService.java
@@ -1,33 +1,0 @@
-package moadong.club.service;
-
-import lombok.AllArgsConstructor;
-import moadong.club.entity.Club;
-import moadong.club.entity.ClubRecruitmentInformation;
-import moadong.club.payload.dto.ClubDetailedResult;
-import moadong.club.payload.response.ClubDetailedResponse;
-import moadong.club.repository.ClubRepository;
-import moadong.global.exception.ErrorCode;
-import moadong.global.exception.RestApiException;
-import org.bson.types.ObjectId;
-import org.springframework.stereotype.Service;
-
-import java.util.Collections;
-import java.util.List;
-
-@Service
-@AllArgsConstructor
-public class ClubDetailedPageService {
-
-    private final ClubRepository clubRepository;
-
-    public ClubDetailedResponse getClubDetailedPage(String clubId) {
-        ObjectId objectId = new ObjectId(clubId);
-        Club club = clubRepository.findClubById(objectId)
-                .orElseThrow(() -> new RestApiException(ErrorCode.CLUB_NOT_FOUND));
-
-        ClubDetailedResult clubDetailedResult = ClubDetailedResult.of(
-                club
-        );
-        return new ClubDetailedResponse(clubDetailedResult);
-    }
-}
